Made the button click change the state of  main Quiz component
By making a chain of function references.

Essentially, Quiz comp. is telling QuizQuestion that whenever click is supposed to be handled, it has to invoke the passed in change Quiz comp. state behavior (showNextQ...).

Now, within the QuizQuestion, it tells the QuizQuestionButton component that 'this is the function you have to execute when click happens.

Inside the QuizQuestionButton component, it gives the actual button a true event handler, and says 'upon this, you execute this function I was given'.

So this sort of makes up an instructions chain down to the actual even handler, which kicks off execution of these instructions.

Since it's only function refs that are being passed in, the actual execution of functions happens within the components, where those functions reside in. (Which makes changing parent parent state possible).

Now the '.bind(this)' - not completely sure what this does. If I had to guess it binds the object's (who's executing the function) instance reference to the function, so that the context wouldn't be lost.